Revert "Kernel: Xiaomi kernel changes for Redmi 5, Redmi 4A, Redmi 5A."
This reverts commit 2dd7cd829e165917bfe384febf21d9727b05fba4.

+++ kernel/sched/sched_avg.c
@@ -107,7 +107,7 @@ EXPORT_SYMBOL(sched_get_nr_running_avg);
  */
 void sched_update_nr_prod(int cpu, long delta, bool inc)
 {
-	s64 diff;
+	int diff;
 	s64 curr_time;
 	unsigned long flags, nr_running;
